Fix a serious bug resulting in corrupted torrents when downloading torrents >4GB on a 32 bit platform.
The problem was the incorrect use of 'size_t' to mark an offset in a file; size_t consists of the maximum length a pointer can point to, which is 4GB on 32 bit platforms (which is why the bug didn't occur on 64 bit platforms). This commit resolves the problem by using 'off_t' when specifying file offsets, which is always >=64 bit (on modern platforms at least) and thus will always yield correct results.

+++ include/tortilla/filemanager.h
@@ -34,10 +34,10 @@ public:
 	void removeFile(File* f);
 
 	//! \brief Write to a file
-	void writeFile(File* f, size_t offset, const void* buf, size_t len);
+	void writeFile(File* f, off_t offset, const void* buf, size_t len);
 
 	//! \brief Read from a file
-	void readFile(File* f, size_t offset, void* buf, size_t len);
+	void readFile(File* f, off_t offset, void* buf, size_t len);
 
 protected:
 	/*! \brief Used to close unused files
