+++ lib/torrent.cc
@@ -892,7 +892,7 @@ Torrent::handleChunk(unsigned int piece, unsigned int offset, uint8_t* buf, size
 	assert(length <= TORRENT_CHUNK_SIZE);
 
 	/* Calculate the absolute position */
-	size_t absolutePos = (size_t)piece * (size_t)pieceLen + offset;
+	off_t absolutePos = (off_t)piece * (off_t)pieceLen + (off_t)offset;
 
 	/* Locate the first file matching this position */
 	unsigned int idx = 0;
@@ -900,7 +900,7 @@ Torrent::handleChunk(unsigned int piece, unsigned int offset, uint8_t* buf, size
 	RLOCK(files);
 	while (idx < files.size()) {
 		if (absolutePos < files[idx]->getLength()) {
-			/* At least a part of the offset to handle resized in this file */
+			/* At least a part of the offset to handle resides in this file */
 			f = files[idx];
 			break;
 		}
